Add split-screen homepage with animated 3D map
- Create split-screen layout with login form on left, 3D map on right
- Add AnimatedMap component with rotating camera around Richmond, VT
- Extract LoginForm as reusable component for homepage and login page
- Add 3D terrain with elevation exaggeration and sky atmosphere
